[#40] create Description component
기존 lazyitem 컴포넌트에 있는 하단의 설명관련 부분을 따로 컴포넌트화
